sync openthread

+++ src/common/message.cc
@@ -509,7 +509,7 @@ bool Message::IsChildPending() const
 {
     bool rval = false;
 
-    for (int i = 0; i < sizeof(msg_child_mask); i++)
+    for (size_t i = 0; i < sizeof(msg_child_mask); i++)
     {
         if (msg_child_mask[i] != 0)
         {
